third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/invalid-meta-directives.html

<!DOCTYPE html>
<html>
<head>
    <title>Invalid directives in meta-delivered CSP</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
	<meta http-equiv="Content-Security-Policy" content="sandbox allow-scripts; frame-ancestors 'none'; report-uri http://foo.test">
</head>
<body>
<script>
test(function () {
    assert_equals(window.location.origin, "http://127.0.0.1:8000");
});
</script>
</body>
</html>
