third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/invalid-meta-directives-expected.txt

CONSOLE ERROR: line 7: Content Security Policies delivered via a <meta> element may not contain the sandbox directive.
CONSOLE ERROR: line 7: Content Security Policies delivered via a <meta> element may not contain the frame-ancestors directive.
CONSOLE ERROR: line 7: Content Security Policies delivered via a <meta> element may not contain the report-uri directive.
This is a testharness.js-based test.
PASS Invalid directives in meta-delivered CSP
Harness: the test ran to completion.
